[Fix] Larastan level 5 errors: Wrong types passed in arguments

+++ app/Game/Services/MoveSubmarineService.php
@@ -31,7 +31,7 @@ class MoveSubmarineService
             ->getDistanceSquaredMovablePerActionPoint();
 
         return new ActionPoints(
-            ceil(
+            (int) ceil(
                 $distanceSquared->getRoot()
                 / $distanceSquaredMovablePerActionPoint->getRoot()
             )
